backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg[binary]==3.2.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
python-dateutil==2.9.0
pytest==8.2.2

